backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.22,<1.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
pydantic>=2.0,<3.0
email-validator>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pytest>=7.0,<9.0
httpx>=0.24,<1.0
python-multipart>=0.0.9
dashscope>=1.15.0
openai>=1.0.0